Update directive.js
Return the custom date object rather than a javascript date object in the checkDate function. Add months in order rather than reversed. Update the original model on date change. Allow for a calculation in the "startYear" attribute.

My Angular skills are limited as I am new to this but the fixes implemented resolve my personal needs - maybe someone smarter can fix this properly.

There is also a bug on line 38 where "!date.month" will evaluate to true in the event that month is 0 (which it will be if January is selected) and that will cause the checkDate function to return false. I will resolve this shortly.